Update README.md
I have a gut feeling that conflicting property descriptors should work like overrides rather than merges. If a stamp is altering the descriptor, it's doing so for a reason. If another stamp is altering it in a different way, it's likely to break the intended behavior of the first anyway, so shouldn't it override?