#23 stamp baseObject parameter update.
This proposal uses an options argument to pass in the `baseObject`, which allows a stamp to have a single options argument through which all composed stamps may access options.

It also suggests less surprising prototype behavior. Instead of ignoring the delegate settings in the stamps (which could break stamp behavior), we can insert a delegate prototype between the original and the instance.